+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Move rails_12factor to a group block
 - Not remove favicon.ico from generated app
 - Set ruby 2.1.0 as default
+- Use jumpup-heroku gem on integration
 
 ### bug fixes